Combine conversation flow into single module
